requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
aiosqlite
asyncpg
langchain
langchain-core
langchain-google-genai
langgraph
chromadb
pydantic
whitenoise
pytest
httpx

